tests: split out toy_server from test_interface.py
